RANDOMWALK-161: rewrite code with interfaces for services
